Started event handling, removed argument checking.
mouse_moved, mouse_dragged, mouse_pressed, and mouse_released are now
supported.

begin() doesn't need to be called with function arguments anymore.

Argument checking was way too slow, so it has been removed be default.
In the future, there will be a debug setting which can be switched on
to enable argument checking.

Small change to the profiler to support function decorators with the
new begin() function.