save_frame() and variable arity to event handlers.
save_frame() is now supported with a variety of image formats.
Currently no other parameters are supported, but quality configurations
will possibly be supported in the future.

Event handlers now do not have to include all of the parameters that can
be passed to them. However, they have a constraint on the names of their
parameters. For example,

    def key_pressed():
        pass

    def key_pressed(key):
        pass

    def key_pressed(x, y):
        pass

    def key_pressed(key, x, y):
    	pass

These are all valid signatures of the key_pressed handler. Condor will
detect which keyword arguments are in the signature and pass in only
those. This prevents naming arguments anything other than key, x, or y,
in the example of key_pressed.