frame_rate() detail() curve_detail() & fixed bugs.
frame_rate(fps=None), when passed no arguments, returns the most recent
frame-rate. When called with an argument, uses it as a frame_rate cap
for future frames.

detail() and curve_detail() are now at least partially supported, with
the default values for both set at 20.

Changed glutils to a class GLUtil, of which Condor has and instance of.
This was done to prepare for the move to modern OpenGL instead of the
glBegin(), glEnd() paradigm.

Other small changes:

- _before_draw() and _prepare_2d() moved to GLUtil.

- condor.profile now has time() to measure a single function call.

- fixed a bug where rect stroke was not drawn properly with different
  rect modes.

- fixed mouse_y() since it was reporting inverted numbers.

- initial mouse coordinates are now (0, 0)