Fixed ellipse stroke bug. Added mouse_down().
mouse_down(*buttons) is a function used to determine if a certain mouse
button is currently pressed. If *buttons is empty, any mouse button will
cause this function to return True. If multiple buttons are passed, if
any of the buttons are pressed mouse_down will return True.

The mouse_dragged handler can now receive x, y coordinates of where the
mouse currently is.

During the last large commit, a portion of the refactoring was done
incorrectly regarding style.detail.